add ocpirun to ocpiremote sandbox / check if gdbserver exist when importing SDK artifacts

Closes #3707 and #3708

See merge request opencpi/opencpi!1172

+++ runtime/hdl-support/xilinx/importSDKartifacts.sh
@@ -71,7 +71,9 @@ getlib "libgcc_s.so*"
 getlib "ld-*.so*"
 getlib "libpthread*.so*"
 mkdir -p $dest/bin
-cp $(dirname $cross)/gdbserver $dest/bin
+if [ -f $(dirname $cross)/gdbserver ]; then
+  cp $(dirname $cross)/gdbserver $dest/bin
+fi
 # While we are here, dump out the details of the compiler
 mkdir -p gen
 ${cross}gcc -Q --help=target > $dest/gcc-target-options
